% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primirTSS_shiny.R
\name{run_primirTSSapp}
\alias{run_primirTSSapp}
\title{Predict TSSs of miRNA using graphical web interface.}
\usage{
run_primirTSSapp()
}
\value{
A graphical interfaces.
}
\description{
A graphical web interface is provided for users to achieve the functions of
\code{\link{find_tss}} and \code{\link{plot_primiRNA}} to intuitively and
conveniently predict putative TSSs of miRNA.
}
\details{
Users can refer documents of the two functions mentioned ABOVE for details.
}
\examples{
run_primirTSSapp()

}
